# docs/openapi.yaml
openapi: 3.0.0
info:
  title: German School Holidays API
  description: |
    API für die Schulferien in Deutschland, filterbar nach Jahr und/oder Bundesland.
    
    ## Versions
    - **V1**: Legacy API endpoints (backward compatible)
    - **V2**: Enhanced API with advanced filtering, search, and statistics
    
  version: 2.0.0
  contact:
    name: Maximilian Leistner
    url: https://maxleistner.de
  license:
    name: Creative Commons Attribution 4.0
    url: https://creativecommons.org/licenses/by/4.0/

servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://ferien-api.maxleistner.de
    description: Production server

tags:
  - name: V1 (Legacy)
    description: Original API endpoints - backward compatible
  - name: V2 (Enhanced)
    description: Enhanced API with advanced filtering and new features
  - name: System
    description: Health checks and system status

paths:
  # V1 Endpoints (Legacy)
  /api/v1/{year}:
    get:
      tags:
        - V1 (Legacy)
      summary: Get all holidays for a specific year (V1)
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: integer
            minimum: 2022
            maximum: 2027
          example: 2024
      responses:
        '200':
          description: List of holidays for the specified year
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Holiday'
        '404':
          description: Year not found
        '500':
          description: Server error

  /api/v1/{year}/{state}:
    get:
      tags:
        - V1 (Legacy)
      summary: Get holidays for a specific year and state (V1)
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: integer
            minimum: 2022
            maximum: 2027
          example: 2024
        - name: state
          in: path
          required: true
          schema:
            type: string
            enum: [BW, BY, BE, BB, HB, HH, HE, MV, NI, NW, RP, SL, SN, ST, SH, TH]
          example: BY
      responses:
        '200':
          description: List of holidays for the specified year and state
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Holiday'
        '404':
          description: No holidays found for the specified criteria

  # V2 Endpoints (Enhanced)
  /api/v2/{year}:
    get:
      tags:
        - V2 (Enhanced)
      summary: Get holidays for a year with advanced filtering (V2)
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: integer
            minimum: 1900
            maximum: 2100
          example: 2024
        - name: from
          in: query
          schema:
            type: string
            format: date
          example: "2024-03-01"
          description: Start date for filtering (YYYY-MM-DD)
        - name: to
          in: query
          schema:
            type: string
            format: date
          example: "2024-08-31"
          description: End date for filtering (YYYY-MM-DD)
        - name: type
          in: query
          schema:
            type: string
          example: "sommerferien,winterferien"
          description: Comma-separated list of holiday types
        - name: states
          in: query
          schema:
            type: string
          example: "BY,BW,BE"
          description: Comma-separated list of state codes
        - name: fields
          in: query
          schema:
            type: string
          example: "start,end,name,stateCode"
          description: Comma-separated list of fields to include
      responses:
        '200':
          description: Filtered list of holidays
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Holiday'
        '400':
          description: Invalid parameters
        '404':
          description: No holidays found matching criteria

  /api/v2/{year}/{state}:
    get:
      tags:
        - V2 (Enhanced)
      summary: Get holidays for a year and state with filtering (V2)
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: integer
          example: 2024
        - name: state
          in: path
          required: true
          schema:
            type: string
          example: BY
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: type
          in: query
          schema:
            type: string
        - name: fields
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Filtered list of holidays for state
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Holiday'

  /api/v2/current:
    get:
      tags:
        - V2 (Enhanced)
      summary: Get currently active holidays
      parameters:
        - name: states
          in: query
          schema:
            type: string
          example: "BY,BW"
        - name: fields
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of currently active holidays
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Holiday'

  /api/v2/next/{days}:
    get:
      tags:
        - V2 (Enhanced)
      summary: Get holidays in the next N days
      parameters:
        - name: days
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 365
          example: 30
        - name: states
          in: query
          schema:
            type: string
        - name: fields
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of upcoming holidays
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Holiday'

  /api/v2/date/{date}:
    get:
      tags:
        - V2 (Enhanced)
      summary: Check if a specific date is a holiday
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
          example: "2024-07-25"
        - name: states
          in: query
          schema:
            type: string
        - name: fields
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Holiday lookup result
          content:
            application/json:
              schema:
                type: object
                properties:
                  date:
                    type: string
                    format: date
                  isHoliday:
                    type: boolean
                  holidays:
                    type: array
                    items:
                      $ref: '#/components/schemas/Holiday'

  /api/v2/search:
    get:
      tags:
        - V2 (Enhanced)
      summary: Search holidays by text query
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          example: "sommer"
          description: Search query for holiday names
        - name: states
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: integer
        - name: fields
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  query:
                    type: string
                  results:
                    type: integer
                  holidays:
                    type: array
                    items:
                      $ref: '#/components/schemas/Holiday'

  /api/v2/stats/{year}:
    get:
      tags:
        - V2 (Enhanced)
      summary: Get holiday statistics for a year
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: integer
          example: 2024
      responses:
        '200':
          description: Holiday statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/YearStats'

  /api/v2/compare/{yearA}/{yearB}:
    get:
      tags:
        - V2 (Enhanced)
      summary: Compare holidays between two years
      parameters:
        - name: yearA
          in: path
          required: true
          schema:
            type: integer
          example: 2024
        - name: yearB
          in: path
          required: true
          schema:
            type: integer
          example: 2025
      responses:
        '200':
          description: Year comparison
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/YearComparison'

  # System Endpoints
  /health:
    get:
      tags:
        - System
      summary: Basic health check
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
                  uptime:
                    type: integer
                  timestamp:
                    type: string
                    format: date-time
                  version:
                    type: string

  /ready:
    get:
      tags:
        - System
      summary: Readiness check for load balancers
      responses:
        '200':
          description: Service is ready
        '503':
          description: Service is not ready

  /status:
    get:
      tags:
        - System
      summary: Comprehensive status page
      responses:
        '200':
          description: Detailed service status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceStatus'

components:
  schemas:
    Holiday:
      type: object
      properties:
        start:
          type: string
          format: date-time
          example: "2024-07-25T00:00Z"
          description: Start date of holiday (ISO 8601)
        end:
          type: string
          format: date-time
          example: "2024-09-08T00:00Z"
          description: End date of holiday (ISO 8601)
        year:
          type: integer
          example: 2024
          description: Year of the holiday
        stateCode:
          type: string
          example: "BY"
          description: German state code
        name:
          type: string
          example: "sommerferien"
          description: Holiday type name in German
        slug:
          type: string
          example: "sommerferien-2024-BY"
          description: Unique identifier for the holiday

    YearStats:
      type: object
      properties:
        year:
          type: integer
        totalHolidays:
          type: integer
        byState:
          type: object
          additionalProperties:
            type: integer
        byType:
          type: object
          additionalProperties:
            type: integer
        averageDuration:
          type: number
        longestHoliday:
          $ref: '#/components/schemas/Holiday'
        shortestHoliday:
          $ref: '#/components/schemas/Holiday'

    YearComparison:
      type: object
      properties:
        yearA:
          type: integer
        yearB:
          type: integer
        totalHolidaysA:
          type: integer
        totalHolidaysB:
          type: integer
        difference:
          type: integer
        byState:
          type: object
        byType:
          type: object

    ServiceStatus:
      type: object
      properties:
        service:
          type: string
        version:
          type: string
        status:
          type: string
        uptime:
          type: integer
        timestamp:
          type: string
          format: date-time
        endpoints:
          type: object
        data:
          type: object
        system:
          type: object

    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string